+++ frontend/src/features/countdown/Countdown.tsx
@@ -1,4 +1,5 @@
-const SECONDS_IN_MINUTE = 60;
+import { SECONDS_IN_MINUTE } from "@/lib/constants";
+
 const MIN_MINUTES_STRING_LENGTH = 2;
 
 type CountdownProps = {
@@ -9,16 +10,15 @@ type CountdownProps = {
 function Countdown({ isHidden, remainingTime, ...props }: CountdownProps) {
   if (isHidden) return;
 
-  const remainingMinutes: number = Math.floor(
-    remainingTime / SECONDS_IN_MINUTE,
+  const minutes = Math.floor(remainingTime / SECONDS_IN_MINUTE);
+  const seconds = String(remainingTime % SECONDS_IN_MINUTE).padStart(
+    MIN_MINUTES_STRING_LENGTH,
+    "0",
   );
-  const remainingSeconds: string = String(
-    remainingTime % SECONDS_IN_MINUTE,
-  ).padStart(MIN_MINUTES_STRING_LENGTH, "0");
 
   return (
     <span {...props}>
-      {remainingMinutes}:{remainingSeconds}
+      {minutes}:{seconds}
     </span>
   );
 }