+++ frontend/src/features/matches/index.ts
@@ -1,3 +1,4 @@
+export * from "./constants";
 export * from "./context/GridSettingsProvider";
 export * from "./images";
 export * from "./services";